simplify shapes & save in geopackage format to reduce file size & reduce number of files
